requirements.txt
Flask==3.0.0
PyMuPDF==1.23.26
google-generativeai==0.8.3
python-dotenv==1.0.0